+++ src/Exceptions/PermissionMustNotBeEmpty.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Spatie\Permission\Exceptions;
-
-use DomainException;
-
-class PermissionMustNotBeEmpty extends DomainException
-{
-}