+++ src/Exceptions/RoleDoesNotExist.php
@@ -6,4 +6,8 @@ use InvalidArgumentException;
 
 class RoleDoesNotExist extends InvalidArgumentException
 {
+    public static function create(string $roleName)
+    {
+        return new static("There is no role named `{$roleName}`.");
+    }
 }