+++ src/PermissionServiceProvider.php
@@ -12,14 +12,14 @@ class PermissionServiceProvider extends ServiceProvider
     public function boot(PermissionRegistrar $permissionLoader)
     {
         $this->publishes([
-            __DIR__.'/../resources/config/laravel-permission.php' => $this->app->configPath().'/'.'laravel-permission.php',
+            __DIR__.'/../resources/config/permission.php' => $this->app->configPath().'/permission.php',
         ], 'config');
 
         if (! class_exists('CreatePermissionTables')) {
             $timestamp = date('Y_m_d_His', time());
 
             $this->publishes([
-                __DIR__.'/../resources/migrations/create_permission_tables.php.stub' => $this->app->databasePath().'/migrations/'.$timestamp.'_create_permission_tables.php',
+                __DIR__.'/../resources/migrations/create_permission_tables.php.stub' => $this->app->databasePath()."/migrations/{$timestamp}_create_permission_tables.php",
             ], 'migrations');
         }
 
@@ -31,8 +31,8 @@ class PermissionServiceProvider extends ServiceProvider
     public function register()
     {
         $this->mergeConfigFrom(
-            __DIR__.'/../resources/config/laravel-permission.php',
-            'laravel-permission'
+            __DIR__.'/../resources/config/permission.php',
+            'permission'
         );
 
         $this->registerBladeExtensions();
@@ -40,7 +40,7 @@ class PermissionServiceProvider extends ServiceProvider
 
     protected function registerModelBindings()
     {
-        $config = $this->app->config['laravel-permission.models'];
+        $config = $this->app->config['permission.models'];
 
         $this->app->bind(PermissionContract::class, $config['permission']);
         $this->app->bind(RoleContract::class, $config['role']);